+++ lib/components/Sound.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import Octavian from 'octavian';
 
-
 const AudioContext = window.AudioContext || window.webkitAudioContext;
 const context = new AudioContext();
 let osc = context.createOscillator();
 let gain = context.createGain();
 const destination = context.destination;
-// var gainNode = context.createGain();
 
 export default class Sound extends React.Component {
   constructor(props) {
@@ -17,12 +15,11 @@ export default class Sound extends React.Component {
     };
   }
 
-
   componentDidMount() {
-    audio.play();
     osc.connect(gain);
     gain.connect(destination);
     osc.start(0);
+    osc.frequency.value = 300;
     gain.gain.value = 1;
   }
