+++ lib/components/TrackColor.js
@@ -8,22 +8,7 @@ export default class TrackColor extends React.Component {
   }
 
   componentDidMount() {
-    this.createVideo();
     this.tracker();
-  }
-
-  createVideo() {
-    navigator.mediaDevices.getUserMedia({ audio: false, video: true })
-    .then((stream) => {
-      const video = document.querySelector('#video');
-      video.src = window.URL.createObjectURL(stream);
-      video.onloadedmetadata = function(e) {
-      video.play();
-    };
-  })
-    .catch((err) => {
-      console.log(err.name + ": " + err.message);
-    });
   }
 
   tracker() {
@@ -35,7 +20,7 @@ export default class TrackColor extends React.Component {
     tracker.on('track', (event) => {
       // context.clearRect(0, 0, canvs.width, canvas.height);
       event.data.forEach((rect) => {
-        context.strokeRect(rect.x, rect.y, rect.width, rect.height);
+        context.strokeRect(rect.x, rect.y, 30, 30);
         context.strokeStyle = '#ADEDE0';
         if (rect) { this.props.onNoteChange(findIndex(rect)); }
       });
@@ -45,7 +30,6 @@ export default class TrackColor extends React.Component {
   render() {
     return (
       <section>
-        <video id="video" height="500" width="800"></video>
         <canvas id="canvas" height="500" width="800"></canvas>
       </section>
     )
