+++ public/index.html
@@ -11,5 +11,6 @@
 <body>
   <div id="application"></div>
   <script src="main.bundle.js"></script>
+  <!-- <script src="webgazer.js" type="text/javascript"></script> -->
 </body>
 </html>
